Move pw validation into Controller
Makes validation more consistent as all other validation is there.

Also changed the tests & 1 use in the GUI code to point to the new location
