+++ project/src/main/Account.java
@@ -20,34 +20,4 @@ public abstract class Account
 	{
 		return username;
 	}
-	
-	/**
-	 * Validates account passwords based on some rules
-	 * TODO: document the rules
-	 * @author krismania
-	 */
-	public static boolean passwordAccepted(String password)
-	{
-		//check length
-		if (password.length() >= 6 && password.length() < 15 && password.matches("[a-zA-Z0-9]+"))
-		{
-			// loop through each character and check if there is (at least) 1 upper case, 1 lower case and 1 number
-			// the following vars are set to true once an occurrence of each is found. -kg
-			boolean upper = false;
-			boolean lower = false;
-			boolean num = false;
-			
-			for (int i = 0; i < password.length(); i++)
-			{
-				// the surrounding if statements prevent extra checks once criteria have been met. -kg
-				if (!upper) if (Character.isUpperCase(password.charAt(i))) upper = true;
-				if (!lower) if (Character.isLowerCase(password.charAt(i))) lower = true;
-				if (!num) if (Character.isDigit(password.charAt(i))) num = true;
-			}
-			
-			// return true if all criteria are met. -kg
-			return upper && lower && num;
-		}
-		return false;
-	}
 }
