+++ project/src/tests/ControllerTest.java
@@ -119,4 +119,110 @@ public class ControllerTest
 		assertTrue(controller.validatePhoneNumber("(03) 9876 5432"));
 	}
 	
+	// test valid passwords. -kg
+	
+	@Test
+	public void testValidPassword1()
+	{		
+		String password = "SomePassword123";
+		assertTrue(controller.passwordAccepted(password));
+	}
+	
+	@Test
+	public void testValidPassword2()
+	{		
+		String password = "5ApwWith5omeNumb3rs!";
+		assertTrue(controller.passwordAccepted(password));
+	}
+	
+	@Test
+	public void testValidPassword3()
+	{		
+		String password = "1234aB";
+		assertTrue(controller.passwordAccepted(password));
+	}
+	
+	@Test
+	public void testValidPassword4()
+	{		
+		String password = "Aaaaa9";
+		assertTrue(controller.passwordAccepted(password));
+	}
+	
+	@Test
+	public void testValidPassword5()
+	{		
+		String password = "1111hhhhhXXXXX";
+		assertTrue(controller.passwordAccepted(password));
+	}
+	
+	
+	// test passwords with invalid length. -kg
+	
+	@Test
+	public void testInvalidLengthPassword1()
+	{		
+		String password = "2Shrt";
+		assertFalse(controller.passwordAccepted(password));
+	}
+	
+	@Test
+	public void testInvalidLengthPassword2()
+	{		
+		String password = "aA1";
+		assertFalse(controller.passwordAccepted(password));
+	}
+
+	@Test
+	public void testInvalidLengthPassword3()
+	{		
+		String password = "1XYab";
+		assertFalse(controller.passwordAccepted(password));
+	}
+
+	@Test
+	public void testInvalidLengthPassword4()
+	{		
+		String password = "!@6Xg";
+		assertFalse(controller.passwordAccepted(password));
+	}
+
+	@Test
+	public void testInvalidLengthPassword5()
+	{		
+		String password = "b7*U";
+		assertFalse(controller.passwordAccepted(password));
+	}
+	
+	
+	// test some specific cases. -kg
+	
+	@Test
+	public void testPasswordNoNumbers()
+	{		
+		String password = "NoNumbersHere";
+		assertFalse(controller.passwordAccepted(password));
+	}
+	
+	@Test
+	public void testPasswordNoUppercase()
+	{		
+		String password = "alllowercasepassword123";
+		assertFalse(controller.passwordAccepted(password));
+	}
+	
+	@Test
+	public void testPasswordNoLowercase()
+	{		
+		String password = "111CAPSLOCK1SSTUCK";
+		assertFalse(controller.passwordAccepted(password));
+	}
+	
+	@Test
+	public void testPasswordNoLetters()
+	{		
+		String password = "123456";
+		assertFalse(controller.passwordAccepted(password));
+	}
+	
 }
